# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateMomentRequest:
    properties:
      content:
        $ref: '#/definitions/dto.MomentContent'
      is_publish:
        description: 是否发布，默认true
        type: boolean
      publish_time:
        allOf:
        - $ref: '#/definitions/utils.JSONTime'
        description: 发布时间（可选，不填则使用创建时间）
    required:
    - content
    type: object
...
  dto.MomentContent:
    properties:
      book:
        additionalProperties: {}
        description: 书籍
        type: object
      images:
        description: 图片列表
        items:
          type: string
        type: array
      link:
        allOf:
        - $ref: '#/definitions/dto.MomentLink'
        description: 外链
      location:
        description: 位置信息
        type: string
      movie:
        additionalProperties: {}
        description: 电影
        type: object
      music:
        allOf:
        - $ref: '#/definitions/dto.MomentMusic'
        description: 音乐（基于MetingJS）
      tags:
        description: 标签
        type: string
      text:
        description: 文本
        type: string
      video:
        allOf:
        - $ref: '#/definitions/dto.MomentVideo'
        description: 视频（本地或在线）
    type: object
  dto.MomentLink:
    properties:
      favicon:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  dto.MomentListResponse:
    properties:
      content:
        $ref: '#/definitions/dto.MomentContent'
      id:
        type: integer
      is_publish:
        type: boolean
      publish_time:
        $ref: '#/definitions/utils.JSONTime'
    type: object
  dto.MomentMusic:
    properties:
      id:
        description: 音乐ID
        type: string
      server:
        description: 音乐平台：netease, tencent, kugou, xiami, baidu
        type: string
      type:
        description: 类型：song, playlist, album, search, artist
        type: string
    type: object
  dto.MomentVideo:
    properties:
      platform:
        description: 平台：bilibili, youtube（本地视频为空）
        type: string
      url:
        description: 视频URL（本地视频或在线视频链接）
        type: string
      video_id:
        description: 视频ID（在线视频的ID，本地视频为空）
        type: string
    type: object
...
  dto.SettingGroupResponse:
    properties:
      group:
        type: string
      settings:
        items:
          $ref: '#/definitions/dto.SettingItemResponse'
        type: array
    type: object
  dto.SettingItemResponse:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  dto.UpdateMomentRequest:
    properties:
      content:
        $ref: '#/definitions/dto.MomentContent'
      is_publish:
        description: 是否发布
        type: boolean
      publish_time:
        allOf:
        - $ref: '#/definitions/utils.JSONTime'
        description: 发布时间（可编辑）
    required:
    - content
    type: object
  dto.UpdateSettingGroupRequest:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    required:
    - settings
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  utils.JSONTime:
    properties:
      time.Time:
        type: string
    type: object
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
...
  /admin/moments:
    get:
      consumes:
      - application/json
      description: 获取所有动态用于管理
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 动态列表（管理）
      tags:
      - 动态管理
    post:
      consumes:
      - application/json
      description: 创建新动态
      parameters:
      - description: 动态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMomentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MomentListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建动态
      tags:
      - 动态管理
  /admin/moments/{id}:
    delete:
      consumes:
      - application/json
      description: 删除动态
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除动态
      tags:
      - 动态管理
    get:
      consumes:
      - application/json
      description: 通过 ID 获取动态详情
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MomentListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 动态详情（管理）
      tags:
      - 动态管理
    put:
      consumes:
      - application/json
      description: 修改动态信息
      parameters:
      - description: 动态 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 动态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMomentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MomentListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新动态
      tags:
      - 动态管理
...
  /admin/settings/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定分组的所有配置项（需要管理员及以上权限）
      parameters:
      - description: 配置分组
        enum:
        - basic
        - blog
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SettingGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取配置分组
      tags:
      - 配置管理
    patch:
      consumes:
      - application/json
      description: 批量更新指定分组的配置项（patch 方式，只更新传入的配置，仅超级管理员可修改）
      parameters:
      - description: 配置分组
        enum:
        - basic
        - blog
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      - description: 配置更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSettingGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新配置分组
      tags:
      - 配置管理
  /admin/settings/ai/mcp-secret/reset:
    put:
      consumes:
      - application/json
      description: 重新生成并返回 MCP Secret（仅超级管理员可修改）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 重置 MCP Secret
      tags:
      - 配置管理
...
  /moments:
    get:
      consumes:
      - application/json
      description: 获取所有公开的动态
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
      summary: 动态列表
      tags:
      - 动态
...
  /settings/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定分组的公开配置项，根据配置项可见性规则筛选
      parameters:
      - description: 配置分组
        enum:
        - basic
        - blog
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取配置分组
      tags:
      - 配置
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"